refactor: price->openapi 도메인 디렉토리명 수정
