+++ src/main/java/hekireki/sanjijiksong/domain/openapi/controller/OpenAPIController.java
@@ -1,12 +1,14 @@
-package hekireki.sanjijiksong.domain.price.controller;
+package hekireki.sanjijiksong.domain.openapi.controller;
 
 
-import hekireki.sanjijiksong.domain.price.dto.ProductPriceResponse;
-import hekireki.sanjijiksong.domain.price.service.PriceService;
-import hekireki.sanjijiksong.domain.price.service.ProductPriceService;
+import hekireki.sanjijiksong.domain.openapi.dto.ProductPriceResponse;
+import hekireki.sanjijiksong.domain.openapi.service.CrawlerService;
+import hekireki.sanjijiksong.domain.openapi.service.PriceService;
+import hekireki.sanjijiksong.domain.openapi.service.ProductPriceService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -15,13 +17,15 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-public class PriceController {
+@RequestMapping("/api/openapi")
+public class OpenAPIController {
     private final PriceService priceService;
     private final ProductPriceService productPriceService;
+    private final CrawlerService crawlerService;
 
 
     // 가격 정보를 가져오는 API 엔드포인트
-    @GetMapping("/api/v1/test/price")
+    @GetMapping("/test/price")
     public ResponseEntity<?> getPrice(@RequestParam(name = "category_code") String categoryCode,
                                       @RequestParam(name = "regday") String regDay) {
         priceService.getPrice(categoryCode, regDay);
@@ -50,5 +54,11 @@ public class PriceController {
         }
         List<ProductPriceResponse> response = productPriceService.getPriceInfo(startDate, endDate, categoryCode, itemCode);
         return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/v1/crawler")
+    public ResponseEntity<?> getCrawler(){
+        crawlerService.crawling();
+        return ResponseEntity.ok("Crawler data fetched successfully");
     }
 }
