+++ src/main/java/hekireki/sanjijiksong/domain/openapi/dto/PriceHistory.java
@@ -1,4 +1,4 @@
-package hekireki.sanjijiksong.domain.price.dto;
+package hekireki.sanjijiksong.domain.openapi.dto;
 
 
 import lombok.Builder;
