+++ src/main/java/hekireki/sanjijiksong/domain/openapi/service/PriceService.java
@@ -1,9 +1,9 @@
-package hekireki.sanjijiksong.domain.price.service;
+package hekireki.sanjijiksong.domain.openapi.service;
 
 
-import hekireki.sanjijiksong.domain.price.Repository.PriceDailyRepository;
-import hekireki.sanjijiksong.domain.price.dto.kamisDailyPrice.KamisDailyResponse;
-import hekireki.sanjijiksong.domain.price.entity.PriceDaily;
+import hekireki.sanjijiksong.domain.openapi.Repository.PriceDailyRepository;
+import hekireki.sanjijiksong.domain.openapi.dto.kamisDailyPrice.KamisDailyResponse;
+import hekireki.sanjijiksong.domain.openapi.entity.PriceDaily;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
