+++ src/main/java/hekireki/sanjijiksong/global/config/SecurityConfig.java
@@ -69,6 +69,7 @@ public class SecurityConfig {
                                 "/api/v1/register"
                         ).permitAll() //해당 url경로는 인증 필요 없음
                         .requestMatchers("/api/v1/admin").hasRole(Role.ADMIN.name())// ADMIN만 접근 가능
+                        .requestMatchers("/api/openapi/**").permitAll()
 //                .requestMatchers().hasRole(Role.BUYER.name())//Buyer만 접근 가능
                         .requestMatchers(HttpMethod.GET, "/api/v1/stores/*").permitAll()
                         .requestMatchers(HttpMethod.POST, "/api/v1/stores/**").hasRole("SELLER")
